Build taggings in redis.
This will let us to an efficient intersection via sinter, which
is not something couch map/reduce is great at.

Also, I don't want to store everything in redis to keep costs low, and
also not worry about data loss from power problems/outages.  Since the
data is just indexing on top of the CouchDB links, it can be flushed and
repopulated pretty fast and easy.